# docs/openapi.json
{"openapi":"3.1.0","info":{"title":"WhatsApp Assistant API","description":"API for AI-powered WhatsApp business assistant with memory and RAG","version":"1.0.0"},"paths":{"/":{"get":{"tags":["Health"],"summary":"Root","description":"Root endpoint with API information.","operationId":"root__get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/health":{"get":{"tags":["Health"],"summary":"Health Check","description":"Check API health and service status.","operationId":"health_check_health_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HealthResponse"}}}}}}},"/upload":{"post":{"tags":["Knowledge Base"],"summary":"Upload Document","description":"Uploads a document/PDF from the user's device and triggers a RAG refresh \nfor the specified business in the background.","operationId":"upload_document_upload_post","requestBody":{"content":{"multipart/form-data":{"schema":{"$ref":"#/components/schemas/Body_upload_document_upload_post"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/chat":{"post":{"tags":["Chat"],"summary":"Chat","description":"Send a message to the assistant and get a response.","operationId":"chat_chat_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ChatRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ChatResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/rag/initialize":{"post":{"tags":["Knowledge Base"],"summary":"Initialize Rag Endpoint","description":"Initialize RAG system for a specific business.","operationId":"initialize_rag_endpoint_rag_initialize_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/InitRAGRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/rag/refresh":{"post":{"tags":["Knowledge Base"],"summary":"Refresh Rag Endpoint","description":"Refresh RAG system for a business (e.g., when knowledge base is updated).","operationId":"refresh_rag_endpoint_rag_refresh_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RefreshRAGRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/profile/{business_id}/{user_id}":{"get":{"tags":["User Profile"],"summary":"Get User Profile","description":"Retrieve a user's profile information.","operationId":"get_user_profile_profile__business_id___user_id__get","parameters":[{"name":"business_id","in":"path","required":true,"schema":{"type":"string","title":"Business Id"}},{"name":"user_id","in":"path","required":true,"schema":{"type":"string","title":"User Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserProfileResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"delete":{"tags":["User Profile"],"summary":"Delete User Profile","description":"Delete a user's profile and conversation history.","operationId":"delete_user_profile_profile__business_id___user_id__delete","parameters":[{"name":"business_id","in":"path","required":true,"schema":{"type":"string","title":"Business Id"}},{"name":"user_id","in":"path","required":true,"schema":{"type":"string","title":"User Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/conversations/{business_id}/{user_id}":{"get":{"tags":["Conversations"],"summary":"Get Conversation History","description":"Retrieve conversation history for a user.","operationId":"get_conversation_history_conversations__business_id___user_id__get","parameters":[{"name":"business_id","in":"path","required":true,"schema":{"type":"string","title":"Business Id"}},{"name":"user_id","in":"path","required":true,"schema":{"type":"string","title":"User Id"}},{"name":"thread_id","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Thread Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}}},"components":{"schemas":{"Body_upload_document_upload_post":{"properties":{"business_id":{"type":"string","title":"Business Id","description":"Unique identifier for the business to assign the document to"},"document_file":{"type":"string","format":"binary","title":"Document File","description":"Document or PDF file to upload and refresh RAG with"}},"type":"object","required":["business_id","document_file"],"title":"Body_upload_document_upload_post"},"ChatRequest":{"properties":{"business_id":{"type":"string","title":"Business Id","description":"Unique identifier for the business"},"user_id":{"type":"string","title":"User Id","description":"Unique identifier for the user/customer"},"message":{"type":"string","title":"Message","description":"User's message"},"thread_id":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Thread Id","description":"Conversation thread ID (auto-generated if not provided)"}},"type":"object","required":["business_id","user_id","message"],"title":"ChatRequest"},"ChatResponse":{"properties":{"response":{"type":"string","title":"Response","description":"Assistant's response"},"thread_id":{"type":"string","title":"Thread Id","description":"Conversation thread ID"},"business_id":{"type":"string","title":"Business Id"},"user_id":{"type":"string","title":"User Id"}},"type":"object","required":["response","thread_id","business_id","user_id"],"title":"ChatResponse"},"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"HealthResponse":{"properties":{"status":{"type":"string","title":"Status"},"database":{"type":"string","title":"Database"},"rag_systems":{"additionalProperties":{"type":"boolean"},"type":"object","title":"Rag Systems"}},"type":"object","required":["status","database","rag_systems"],"title":"HealthResponse"},"InitRAGRequest":{"properties":{"business_id":{"type":"string","title":"Business Id","description":"Business ID to initialize RAG for"},"doc_path":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Doc Path","description":"Path to knowledge base document (CSV/PDF/DOCX)"}},"type":"object","required":["business_id"],"title":"InitRAGRequest"},"RefreshRAGRequest":{"properties":{"business_id":{"type":"string","title":"Business Id","description":"Business ID to refresh RAG for"},"doc_path":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Doc Path","description":"Path to updated knowledge base document"}},"type":"object","required":["business_id"],"title":"RefreshRAGRequest"},"UserProfileResponse":{"properties":{"business_id":{"type":"string","title":"Business Id"},"user_id":{"type":"string","title":"User Id"},"name":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Name"},"location":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Location"},"address":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Address"},"cart":{"anyOf":[{"items":{"type":"string"},"type":"array"},{"type":"null"}],"title":"Cart"},"human_active":{"anyOf":[{"type":"boolean"},{"type":"null"}],"title":"Human Active"}},"type":"object","required":["business_id","user_id"],"title":"UserProfileResponse"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"}}}}